Разделяет утилитарные функции на общие и узкоспециальные
